pyproject.toml
[project]
name = "orion-agent"
version = "0.1.0"
description = "Production-ready AI customer support agent for e-commerce, built with LangGraph, Qdrant, and Supabase"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastembed>=0.7.4",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "langchain-core>=1.2.20",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai",
    "langgraph>=1.1.3",
    "langgraph-checkpoint-sqlite",
    "langsmith>=0.7.22",
    "ollama>=0.6.1",
    "pandas>=1.3.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "ragas",
    "sqlalchemy>=2.0.48",
    "sqlparse>=0.5.5",
    "streamlit>=1.35.0",
]

[project.scripts]
chunker = "ingestion.chunker:main"

[dependency-groups]
dev = [
    "pytest",
    "ruff>=0.15.12",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
